assessmentSectionView: moved Descriptions to its own component
